requirements.txt
Flask==1.1.2
torch==1.3.1
numpy==1.18.1
pandas==1.0.4
Pillow==7.1.2
